Authentication Login for User and host
